Support to consume t function outside React
