+++ examples/without-loader/pages/404.js
@@ -1,6 +1,5 @@
 import useTranslation from 'next-translate/useTranslation'
 import loadNamespaces from 'next-translate/loadNamespaces'
-import i18nConfig from '../i18n'
 
 export default function Error404() {
   const { t, lang } = useTranslation()
@@ -14,7 +13,6 @@ export default function Error404() {
 export async function getStaticProps(ctx) {
   return {
     props: await loadNamespaces({
-      ...i18nConfig,
       ...ctx,
       pathname: '/404',
     }),
