+++ examples/without-loader/pages/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import Link from 'next/link'
 import useTranslation from 'next-translate/useTranslation'
 import loadNamespaces from 'next-translate/loadNamespaces'
-import i18nConfig from '../i18n'
 import Header from '../components/header'
 
 export default function Home() {
@@ -24,7 +23,6 @@ export default function Home() {
 export async function getStaticProps(ctx) {
   return {
     props: await loadNamespaces({
-      ...i18nConfig,
       ...ctx,
       pathname: '/',
     }),
