+++ src/useTranslation.tsx
@@ -1,18 +1,12 @@
 import { useContext } from 'react'
 import { I18n } from '.'
+import wrapTWithDefaultNs from './wrapTWithDefaultNs'
 import I18nContext from './_context'
 
 export default function useTranslation(defaultNs?: string) {
   const ctx = useContext(I18nContext)
-
-  if (typeof defaultNs !== 'string') return ctx
-
-  // Use default namespace if namespace is missing
-  function t(key = '', query, options) {
-    let k = Array.isArray(key) ? key[0] : key
-    if (!k.includes(':')) k = `${defaultNs}:${k}`
-    return ctx.t(k, query, options)
-  }
-
-  return { ...ctx, t } as I18n
+  return {
+    ...ctx,
+    t: wrapTWithDefaultNs(ctx.t, defaultNs),
+  } as I18n
 }
